Updated
I made adjustments to each of the pages including updating the image src's, updated the nav bar, did some image adjustments, added buttons to the about page, updated the buttons for the hobbies page. Unfortunately, I did have to drop the button a:visted since that was not working with the updated buttons.